add login page and token check mechanism
