+++ src/router/index.js
@@ -1,6 +1,7 @@
 import Vue from "vue";
 import VueRouter from "vue-router";
 import HomePage from "../views/HomePage.vue";
+import LoginPage from "../views/LoginPage.vue";
 import RoomPage from "../views/RoomPage.vue";
 import IssuePage from "../views/IssuePage.vue";
 import NotFoundPage from "../views/NotFoundPage.vue";
@@ -12,6 +13,11 @@ const routes = [
     path: "/",
     name: "home",
     component: HomePage,
+  },
+  {
+    path: "/login",
+    name: "login",
+    component: LoginPage,
   },
   {
     path: "/notFound",
