+++ src/utils/constants.js
@@ -1,4 +1,5 @@
 export const STORAGE_UID = "planning-poker-user-id";
+export const STORAGE_TOKEN = "planning-poker-token";
 export const ROLES = {
   USER: "user",
   ADMIN: "admin",
